Remove journald logging in favour of stdout
The result is the same (stdout is logged to journald), while removing the systemd library dependency, which is troublesome for local development as it requires building.

This commit also removes the failsafe for sdnotify being absent, as it should always be present (in d/control and requirements.txt). When not running in systemd, `SystemdNotifier.notify` is a NOOP.